requirements.txt
# Core dependencies
ipython
jupyterlab
loguru
matplotlib
notebook
numpy
pandas
pip
python-dotenv
ruff
scikit-learn
tqdm
typer
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=9.0.0
albumentations>=1.3.0
opencv-python>=4.8.0
onnx>=1.14.0
onnxruntime>=1.15.0
gradio>=3.40.0
tensorboard>=2.13.0
seaborn>=0.12.0
pyyaml>=6.0
wandb>=0.15.0

-e .
